process lines in parallel
